Refactor: group store init and middleware in a directory
